smoothing updates

+++ assembler/src/debruijn/utils.hpp
@@ -1512,9 +1512,9 @@ public:
 
 
     double CountWeight(int x) const {
-        size_t xx = insert_size - left_x + x - 1;
+        int xx = insert_size - left_x + x - 1;
         
-        if (!(xx >= 0 && xx < new_hist.size())) return 0.;
+        if (!(xx >= 0 && xx < (int) new_hist.size())) return 0.;
         //cout << x << " " << xx << " " << new_hist[xx] << endl;
         VERIFY(math::le(new_hist[xx], 1.));
         return new_hist[xx];
